pyproject.toml
[project]
name = "massage-app"
version = "0.1.0"
description = "Gmail IMAP IDLE to MQTT Monitor with popup notifications"
readme = "README_IMPROVED_SERVICE.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "paho-mqtt>=1.6.0",
    "imapclient>=2.1.0",
    "backoff>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]